sync with develop

+++ integtest/multi_output_file_test.py
@@ -4,29 +4,49 @@ import re
 import copy
 import urllib.request
 
-import dfmodules.data_file_checks as data_file_checks
+import integrationtest.data_file_checks as data_file_checks
 import integrationtest.log_file_checks as log_file_checks
 import integrationtest.config_file_gen as config_file_gen
-import dfmodules.integtest_file_gen as integtest_file_gen
+import integrationtest.dro_map_gen as dro_map_gen
 
 # Values that help determine the running conditions
 number_of_data_producers=2
 number_of_readout_apps=3
-data_rate_slowdown_factor=10
+data_rate_slowdown_factor=1
 
 # Default values for validation parameters
 expected_number_of_data_files=4
 check_for_logfile_errors=True
-wib1_frag_hsi_trig_params={"fragment_type_description": "WIB", 
+wib1_frag_hsi_trig_params={"fragment_type_description": "WIB",
                            "fragment_type": "ProtoWIB",
                            "hdf5_source_subsystem": "Detector_Readout",
                            "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
-                           "min_size_bytes": 37192, "max_size_bytes": 185960}
+                           "min_size_bytes": 37192, "max_size_bytes": 186136}
 wib1_frag_multi_trig_params={"fragment_type_description": "WIB",
                              "fragment_type": "ProtoWIB",
                              "hdf5_source_subsystem": "Detector_Readout",
                              "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
                              "min_size_bytes": 72, "max_size_bytes": 270000}
+wib2_frag_hsi_trig_params={"fragment_type_description": "WIB",
+                           "fragment_type": "WIB",
+                           "hdf5_source_subsystem": "Detector_Readout",
+                           "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
+                           "min_size_bytes": 29808, "max_size_bytes": 30280}
+wib2_frag_multi_trig_params={"fragment_type_description": "WIB",
+                             "fragment_type": "WIB",
+                             "hdf5_source_subsystem": "Detector_Readout",
+                             "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
+                             "min_size_bytes": 72, "max_size_bytes": 54000}
+wibeth_frag_hsi_trig_params={"fragment_type_description": "WIBEth",
+                  "fragment_type": "WIBEth",
+                  "hdf5_source_subsystem": "Detector_Readout",
+                  "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
+                  "min_size_bytes": 187272, "max_size_bytes": 194472}
+wibeth_frag_multi_trig_params={"fragment_type_description": "WIBEth",
+                  "fragment_type": "WIBEth",
+                  "hdf5_source_subsystem": "Detector_Readout",
+                  "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
+                  "min_size_bytes": 72, "max_size_bytes": 194472}
 triggercandidate_frag_params={"fragment_type_description": "Trigger Candidate",
                               "fragment_type": "Trigger_Candidate",
                               "hdf5_source_subsystem": "Trigger",
@@ -40,13 +60,13 @@ triggeractivity_frag_params={"fragment_type_description": "Trigger Activity",
 triggertp_frag_params={"fragment_type_description": "Trigger with TPs",
                        "fragment_type": "Trigger_Primitive",
                        "hdf5_source_subsystem": "Trigger",
-                       "expected_fragment_count": ((number_of_data_producers*number_of_readout_apps)),
+                       "expected_fragment_count": (2*number_of_readout_apps),
                        "min_size_bytes": 72, "max_size_bytes": 16000}
 hsi_frag_params ={"fragment_type_description": "HSI",
                              "fragment_type": "Hardware_Signal",
                              "hdf5_source_subsystem": "HW_Signals_Interface",
                              "expected_fragment_count": 1,
-                             "min_size_bytes": 72, "max_size_bytes": 96}
+                             "min_size_bytes": 72, "max_size_bytes": 100}
 ignored_logfile_problems={"trigger": ["zipped_tpset_q: Unable to push within timeout period"]}
 
 # The next three variable declarations *must* be present as globals in the test
@@ -54,30 +74,61 @@ ignored_logfile_problems={"trigger": ["zipped_tpset_q: Unable to push within tim
 # to run the config generation and nanorc
 
 # The name of the python module for the config generation
-confgen_name="daqconf_multiru_gen"
+confgen_name="fddaqconf_gen"
 # The arguments to pass to the config generator, excluding the json
 # output directory (the test framework handles that)
-hardware_map_contents = integtest_file_gen.generate_hwmap_file(number_of_data_producers, number_of_readout_apps)
+dro_map_contents = dro_map_gen.generate_dromap_contents(number_of_data_producers, number_of_readout_apps)
 
 conf_dict = config_file_gen.get_default_config_dict()
-try:
-  urllib.request.urlopen('http://localhost:5000').status
-  conf_dict["boot"]["use_connectivity_service"] = True
-except:
-  conf_dict["boot"]["use_connectivity_service"] = False
-conf_dict["readout"]["data_rate_slowdown_factor"] = data_rate_slowdown_factor
+conf_dict["daq_common"]["data_rate_slowdown_factor"] = data_rate_slowdown_factor
 conf_dict["readout"]["latency_buffer_size"] = 200000
-conf_dict["trigger"]["trigger_rate_hz"] = 10
-conf_dict["trigger"]["trigger_window_before_ticks"] = 5000
-conf_dict["trigger"]["trigger_window_after_ticks"] = 5000
+conf_dict["detector"]["op_env"] = "integtest"
+#conf_dict["readout"]["default_data_file"] = "asset://?label=DuneWIB&subsystem=readout" # DuneWIB
+conf_dict["readout"]["default_data_file"] = "asset://?checksum=e96fd6efd3f98a9a3bfaba32975b476e" # WIBEth
+conf_dict["detector"]["clock_speed_hz"] = 62500000 # DuneWIB/WIBEth
+conf_dict["readout"]["use_fake_cards"] = True
+conf_dict["readout"]["emulated_data_times_start_with_now"] = True
+conf_dict["hsi"]["random_trigger_rate_hz"] = 10
+conf_dict["trigger"]["trigger_window_before_ticks"] = 52000
+conf_dict["trigger"]["trigger_window_after_ticks"] = 1000
 conf_dict["dataflow"]["apps"][0]["max_file_size"] = 1074000000
 
 swtpg_conf = copy.deepcopy(conf_dict)
-swtpg_conf["readout"]["enable_software_tpg"] = True
-swtpg_conf["dataflow"]["apps"][0]["token_count"] = max(10, 3*number_of_data_producers*number_of_readout_apps)
+swtpg_conf["readout"]["generate_periodic_adc_pattern"] = True
+swtpg_conf["readout"]["emulated_TP_rate_per_ch"] = 1
+swtpg_conf["readout"]["enable_tpg"] = True
+swtpg_conf["readout"]["tpg_threshold"] = 500
+swtpg_conf["readout"]["tpg_algorithm"] = "SimpleThreshold"
+swtpg_conf["readout"]["default_data_file"] = "asset://?checksum=dd156b4895f1b06a06b6ff38e37bd798" # WIBEth All Zeros
+swtpg_conf["trigger"]["trigger_activity_plugin"] = ["TriggerActivityMakerPrescalePlugin"]
+swtpg_conf["trigger"]["trigger_activity_config"] = [ {"prescale": 300} ]
+swtpg_conf["trigger"]["trigger_candidate_plugin"] = ["TriggerCandidateMakerPrescalePlugin"]
+swtpg_conf["trigger"]["trigger_candidate_config"] = [ {"prescale": 100} ]
+swtpg_conf["trigger"]["mlt_merge_overlapping_tcs"] = False
+swtpg_conf["dataflow"]["token_count"] = max(10, 6*number_of_data_producers*number_of_readout_apps)
 
-confgen_arguments={"WIB1_System": conf_dict,
-                   "Software_TPG_System": swtpg_conf,
+multiout_conf = copy.deepcopy(conf_dict)
+multiout_conf["dataflow"]["apps"][0]["output_paths"] = [".", "."]
+multiout_conf["dataflow"]["apps"][0]["max_file_size"] = 4*1024*1024*1024
+
+multiout_tpg_conf = copy.deepcopy(multiout_conf)
+multiout_tpg_conf["readout"]["generate_periodic_adc_pattern"] = True
+multiout_tpg_conf["readout"]["emulated_TP_rate_per_ch"] = 1
+multiout_tpg_conf["readout"]["enable_tpg"] = True
+multiout_tpg_conf["readout"]["tpg_threshold"] = 500
+multiout_tpg_conf["readout"]["tpg_algorithm"] = "SimpleThreshold"
+multiout_tpg_conf["readout"]["default_data_file"] = "asset://?checksum=dd156b4895f1b06a06b6ff38e37bd798" # WIBEth All Zeros
+multiout_tpg_conf["trigger"]["trigger_activity_plugin"] = ["TriggerActivityMakerPrescalePlugin"]
+multiout_tpg_conf["trigger"]["trigger_activity_config"] = [ {"prescale": 300} ]
+multiout_tpg_conf["trigger"]["trigger_candidate_plugin"] = ["TriggerCandidateMakerPrescalePlugin"]
+multiout_tpg_conf["trigger"]["trigger_candidate_config"] = [ {"prescale": 100} ]
+multiout_tpg_conf["trigger"]["mlt_merge_overlapping_tcs"] = False
+multiout_tpg_conf["dataflow"]["token_count"] = max(10, 6*number_of_data_producers*number_of_readout_apps)
+
+confgen_arguments={"WIBEth_System (Rollover files)": conf_dict,
+                   "Software_TPG_System (Rollover files)": swtpg_conf,
+                   "WIBEth_System (Multiple outputs)": multiout_conf,
+                   "Software_TPG_System (Multiple outputs)": multiout_tpg_conf
                   }
 
 # The commands to run in nanorc, as a list
@@ -105,13 +156,17 @@ def test_log_files(run_nanorc):
 def test_data_files(run_nanorc):
     local_file_count=expected_number_of_data_files
     fragment_check_list=[triggercandidate_frag_params, hsi_frag_params]
-    if "enable_software_tpg" in run_nanorc.confgen_config["readout"].keys() and run_nanorc.confgen_config["readout"]["enable_software_tpg"]:
-        local_file_count=5
-        fragment_check_list.append(wib1_frag_multi_trig_params)
+    if "enable_tpg" in run_nanorc.confgen_config["readout"].keys() and run_nanorc.confgen_config["readout"]["enable_tpg"]:
+        local_file_count=4  # 5
+        #fragment_check_list.append(wib1_frag_multi_trig_params) # ProtoWIB
+        #fragment_check_list.append(wib2_frag_multi_trig_params) # DuneWIB
+        fragment_check_list.append(wibeth_frag_multi_trig_params) # WIBEth
         fragment_check_list.append(triggertp_frag_params)
         fragment_check_list.append(triggeractivity_frag_params)
     else:
-        fragment_check_list.append(wib1_frag_hsi_trig_params)
+        #fragment_check_list.append(wib1_frag_hsi_trig_params) # ProtoWIB
+        #fragment_check_list.append(wib2_frag_hsi_trig_params) # DuneWIB
+        fragment_check_list.append(wibeth_frag_hsi_trig_params) # WIBEth
 
     # Run some tests on the output data file
     assert len(run_nanorc.data_files)==local_file_count
@@ -125,9 +180,6 @@ def test_data_files(run_nanorc):
             assert data_file_checks.check_fragment_sizes(data_file, fragment_check_list[jdx])
 
 def test_cleanup(run_nanorc):
-    print("============================================")
-    print("Listing the hdf5 files before deleting them:")
-    print("============================================")
     pathlist_string=""
     filelist_string=""
     for data_file in run_nanorc.data_files:
@@ -135,13 +187,18 @@ def test_cleanup(run_nanorc):
         if str(data_file.parent) not in pathlist_string:
             pathlist_string += " " + str(data_file.parent)
 
-    os.system(f"df -h {pathlist_string}")
-    print("--------------------")
-    os.system(f"ls -alF {filelist_string}");
+    if pathlist_string and filelist_string:
+        print("============================================")
+        print("Listing the hdf5 files before deleting them:")
+        print("============================================")
 
-    for data_file in run_nanorc.data_files:
-        data_file.unlink()
+        os.system(f"df -h {pathlist_string}")
+        print("--------------------")
+        os.system(f"ls -alF {filelist_string}");
 
-    print("--------------------")
-    os.system(f"df -h {pathlist_string}")
-    print("============================================")
+        for data_file in run_nanorc.data_files:
+            data_file.unlink()
+
+        print("--------------------")
+        os.system(f"df -h {pathlist_string}")
+        print("============================================")
